Refactored the display side
Good start on changing the display stuff to more logical urls and more
extensible code. This all started with trying to add in displaying
albums instead of songs when an artist is selected, and now allows for
much more changes or additions as necessary in the future.

The other actions (play, etc) still need to be updated.

Other todo: refactor storage and api files. Storage needs to not do this
return a list with magical indicies nonsense.